+++ src/js/ECharts.js
@@ -3,12 +3,9 @@
  * Copyright (C) 2019 Future Internet Consulting and Development Solutions S.L. All Rights Reserved.
  *
  */
+/* globals echarts, ResizeObserver */
 
-/* exported ECharts */
-/* globals echarts */
-
-
-var ECharts = (function () {
+(function () {
 
     "use strict";
 
@@ -16,74 +13,140 @@ var ECharts = (function () {
     // CLASS DEFINITION
     // =========================================================================
 
-    var echart;
+    class Widget {
+        constructor(MashupPlatform, shadowDOM, _) {
+            this.MashupPlatform = MashupPlatform;
+            this.shadowDOM = shadowDOM;
 
+            // Not used, but needed to avoid errors
+            this.MashupPlatform.prefs.registerCallback(function (_) {});
 
-    var eCharts = function eCharts() {
+            this.eCharts();
+        }
 
-        var container = document.getElementById('echartContainer');
-        echart = echarts.init(container);
-        // New ECharts options handler
-        MashupPlatform.wiring.registerCallback("echarts_options", loadChart);
+        eCharts() {
+            var container = this.shadowDOM.getElementById('echartContainer');
+            this.echart = echarts.init(container);
+            // New ECharts options handler
+            this.MashupPlatform.wiring.registerCallback("echarts_options", this.loadChart.bind(this));
 
-        // Resize handler
-        window.addEventListener("resize",() => {
-            if (echart != null) {
-                echart.resize();
+            // Resize handler
+            if ('ResizeObserver' in window) {
+                this.resizeObserver = new ResizeObserver(() => {
+                    if (this.echart != null) {
+                        this.echart.resize();
+                    }
+                });
+
+                this.resizeObserver.observe(container);
+            } else {
+                this.MashupPlatform.widget.context.registerCallback(function (newValues) {
+                    if ("heightInPixels" in newValues || "widthInPixels" in newValues) {
+                        this.echart.resize();
+                    }
+                }.bind(this));
             }
-        });
 
-        /* test-code */
-        window.loadChart = loadChart;
-        /* end-test-code */
-    };
+            this.echart.on('mouseover', (event) => {
+                this.MashupPlatform.wiring.pushEvent('hover', this.createSingleDataEvent('hover', event));
+            });
 
-    var loadChart = function loadChart(data) {
-        if (data == null) {
-            // Load Empty chart
-            echart.clear();
-            echart.showLoading();
+            this.echart.on('click', (event) => {
+                this.MashupPlatform.wiring.pushEvent('click', this.createSingleDataEvent('click', event));
+            });
 
-            var msgOption = {
-                title: {
-                    show: true,
-                    textStyle: {
-                        color: 'grey',
-                        fontSize: 20
+            this.echart.on('dblclick', (event) => {
+                this.MashupPlatform.wiring.pushEvent('dblclick', this.createSingleDataEvent('dblclick', event));
+            });
+
+            this.echart.on('highlight', (event) => {
+                const dataObjs = [];
+                event.batch.forEach((item) => {
+                    dataObjs.push(this.createDataObj(item.seriesIndex, item.dataIndex));
+                });
+                const eventData = {
+                    event: 'highlight',
+                    batch: dataObjs
+                }
+
+                this.MashupPlatform.wiring.pushEvent('highlight', eventData);
+            });
+        }
+
+        loadChart(data) {
+            if (data == null) {
+                // Load Empty chart
+                this.echart.clear();
+                this.echart.showLoading();
+
+                var msgOption = {
+                    title: {
+                        show: true,
+                        textStyle: {
+                            color: 'grey',
+                            fontSize: 20
+                        },
+                        text: "No Data",
+                        left: 'center',
+                        top: 'center'
                     },
-                    text: "No Data",
-                    left: 'center',
-                    top: 'center'
-                },
-                xAxis: {
-                    show: false
-                },
-                yAxis: {
-                    show: false
-                },
-                series: []
-            };
+                    xAxis: {
+                        show: false
+                    },
+                    yAxis: {
+                        show: false
+                    },
+                    series: []
+                };
 
-            echart.setOption(msgOption);
-            echart.hideLoading();
-            return;
-        }
-
-        // Load new chart
-        echart.clear();
-        echart.showLoading();
-
-        if (data && typeof data === "object") {
-            try {
-                echart.setOption(data, true);
-            } catch (e) {
-                MashupPlatform.widget.log("Error loading the new options in ECharts: " + e, MashupPlatform.log.ERROR);
+                this.echart.setOption(msgOption);
+                this.echart.hideLoading();
+                return;
             }
-        } else {
-            MashupPlatform.widget.log("Invalid ECharts options. Should be a JSON object", MashupPlatform.log.ERROR);
-        }
-        echart.hideLoading();
-    };
 
-    return eCharts;
+            // Load new chart
+            this.echart.clear();
+            this.echart.showLoading();
+
+            if (data && typeof data === "string") {
+                try {
+                    data = JSON.parse(data);
+                } catch (e) {
+                    this.MashupPlatform.widget.log("Error parsing the ECharts options: " + e, this.MashupPlatform.log.WARN);
+                }
+            }
+
+            if (data && typeof data === "object") {
+                try {
+                    this.echart.setOption(data, !this.MashupPlatform.prefs.get("merge"));
+                } catch (e) {
+                    this.MashupPlatform.widget.log("Error loading the new options in ECharts: " + e, this.MashupPlatform.log.ERROR);
+                }
+            } else {
+                this.MashupPlatform.widget.log("Invalid ECharts options. Should be a JSON object", this.MashupPlatform.log.ERROR);
+            }
+            this.echart.hideLoading();
+        }
+
+        createDataObj(seriesIndex, dataIndex) {
+            const data = this.echart.getOption().series[seriesIndex].data[dataIndex];
+            return {
+                seriesIndex: seriesIndex,
+                dataIndex: dataIndex,
+                data: data
+            };
+        }
+
+        createSingleDataEvent(name, event) {
+            const dataObj = this.createDataObj(event.seriesIndex, event.dataIndex);
+            const eventData = {
+                event: name,
+                batch: [dataObj]
+            }
+            return eventData;
+        }
+    }
+
+    window.FICODES_ECharts = Widget;
+
 })();